+++ src/main/java/pl/tw/ksiegowosc/mapper/AllegroInvoiceMappingContext.java
@@ -10,5 +10,7 @@ public record AllegroInvoiceMappingContext(
         String invoiceNo,
         LocalDate docDate,
         List<MeritTaxDto> taxes,
-        String uomName) {
+        String uomName,
+        String accountName,
+        String sellerLogin) {
 }